chore: update datafusion (#1114)
# Description

A simple maintenance PR to update datafusion to the latest version. 

# Related Issue(s)
<!---
For example:

- closes #106
--->

# Documentation

<!---
Share links to useful documentation
--->

+++ rust/src/action/parquet_read/mod.rs
@@ -268,15 +268,15 @@ fn primitive_parquet_field_to_json_value(field: &Field) -> Result<serde_json::Va
     }
 }
 
-fn convert_timestamp_millis_to_string(value: u64) -> Result<String, &'static str> {
+fn convert_timestamp_millis_to_string(value: i64) -> Result<String, &'static str> {
     let dt = Utc
-        .timestamp_opt((value / 1000) as i64, ((value % 1000) * 1000000) as u32)
+        .timestamp_opt(value / 1000, ((value % 1000) * 1000000) as u32)
         .single()
         .ok_or("Value out of bounds")?;
     Ok(dt.to_rfc3339_opts(SecondsFormat::Millis, true))
 }
 
-fn convert_date_to_string(value: u32) -> Result<String, &'static str> {
+fn convert_date_to_string(value: i32) -> Result<String, &'static str> {
     static NUM_SECONDS_IN_DAY: i64 = 60 * 60 * 24;
     let dt = Utc
         .timestamp_opt(value as i64 * NUM_SECONDS_IN_DAY, 0)
